CColliderTriangle Collision Detection
Collision detection for the CColliderTriangle type collider has been
implemented although there are still a few bugs to work out as
collisions are not always detected properlly. Some other minor changes
to the code have been made as well but do not effect functionality

+++ PaTFramework/source/CD3DGraphics.h
@@ -46,13 +46,13 @@ public:
 	ID3D11Device*			Device() { return m_pDXDevice; }
 	ID3D11DeviceContext*	Context() { return m_pDXContext; }
 	unsigned int GetClientWidth() {
-		if (m_cTarget < 0 || m_cTarget >= m_pRenderTarget.size()) return 0; 
-		else if (!m_pRenderTarget[m_cTarget]->m_pWnd) return 0; 
+		if (m_cTarget < 0 || m_cTarget >= (int)m_pRenderTarget.size()) return (UINT)0; 
+		else if (!m_pRenderTarget[m_cTarget]->m_pWnd) return (UINT)0; 
 		else return m_pRenderTarget[m_cTarget]->m_pWnd->GetWidth();
 	}
 	unsigned int GetClientHeight() { 
-		if (m_cTarget < 0 || m_cTarget >= m_pRenderTarget.size()) return 0; 
-		else if (!m_pRenderTarget[m_cTarget]->m_pWnd) return 0; 
+		if (m_cTarget < 0 || m_cTarget >= (int)m_pRenderTarget.size()) return (UINT)0; 
+		else if (!m_pRenderTarget[m_cTarget]->m_pWnd) return (UINT)0; 
 		else return m_pRenderTarget[m_cTarget]->m_pWnd->GetHeight();
 	}
